lessons/05/code/configuration/gemini_open_ai.py
from dotenv import load_dotenv
from openai import OpenAI
import os

load_dotenv()

API_KEY = os.getenv("GEMINI_API_KEY")
BASE_URL = "https://generativelanguage.googleapis.com/v1beta/openai/"

client = OpenAI(api_key=API_KEY, base_url=BASE_URL)

response = client.chat.completions.create(
    model="gemini-2.5-flash-lite", messages=[{"role": "user", "content": "Hey there"}]
)


print(response.choices[0].message.content)
